+++ controllers/slice/slicerouter.go
@@ -30,7 +30,7 @@ import (
 
 	"github.com/kubeslice/worker-operator/controllers"
 	"github.com/kubeslice/worker-operator/pkg/logger"
-	webhook "github.com/kubeslice/worker-operator/pkg/webhook/deploy"
+	webhook "github.com/kubeslice/worker-operator/pkg/webhook/pod"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
