+++ pkg/webhook/pod/webhook_utils.go
@@ -1,4 +1,4 @@
-package deploy
+package pod
 
 import (
 	"context"
